backend/src/requirements.txt
annotated-types==0.5.0
async-timeout==4.0.3
Authlib==1.3.1
blinker==1.8.2
cachelib==0.13.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.3
coverage==7.3.1
cryptography==42.0.8
Flask==2.2.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.2
flask-oidc==2.2.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
JSON-log-formatter==1.0
MarkupSafe==2.1.2
msgspec==0.18.6
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.6
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pytest==7.4.2
python-dotenv==1.0.0
redis==5.0.6
requests==2.31.0
SQLAlchemy==2.0.13
typing_extensions==4.8.0
urllib3==2.0.5
waitress==2.1.2
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0